start setting up bin scrupts
