+++ package.json
@@ -13,6 +13,10 @@
     "prepack": "pinst --disable",
     "postpack": "pinst --enable"
   },
+  "bin": {
+    "@entropyxyz/cli": "./src/bin/cli.sh",
+    "entropy": "./src/bin/cli.sh"
+  },
   "repository": {
     "type": "git",
     "url": "git+https://github.com/entropyxyz/cli.git"
@@ -36,6 +40,7 @@
     "dotenv": "^16.4.1",
     "ethers": "^5.7.2",
     "inquirer": "8.0.0",
+    "minimist": "^1.2.8",
     "ts-node": "^10.9.2",
     "typescript": "^4.8.4",
     "viem": "^2.7.8",
