+++ src/bin/entropy.sh
@@ -1,0 +1,3 @@
+#! /usr/bin/env bash
+
+npx ts-node ./src/index.ts
